feat(backend): Authorization for session photos

+++ backend/src/session_photo/session_photo.service.ts
@@ -3,10 +3,50 @@ import { GetSessionPhotoDto } from './dto/get-session_photo.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateSessionPhotoDto } from './dto/create-session_photo.dto';
 import { UpdateSessionPhotoDto } from './dto/update-session.dto';
+import { CurrentUserProvider } from 'src/current-user/current-user.provider';
+import {
+  ForbiddenException,
+  NotFoundException,
+} from 'src/exception/service-exception';
+import { SessionPhoto } from '@prisma/client';
 
 @Injectable()
 export class SessionPhotoService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private currentUserProvider: CurrentUserProvider,
+  ) {}
+
+  private get userId(): string {
+    return this.currentUserProvider.user.userId;
+  }
+
+  private async authorize(id: string): Promise<SessionPhoto> {
+    const sessionPhotoWithProject = await this.prisma.sessionPhoto.findUnique({
+      where: {
+        id: id,
+      },
+      include: {
+        session: {
+          include: {
+            project: true,
+          },
+        },
+      },
+    });
+
+    if (!sessionPhotoWithProject) {
+      throw new NotFoundException('Project not found');
+    }
+
+    if (sessionPhotoWithProject.session.project.userId != this.userId) {
+      throw new ForbiddenException('');
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { session, ...sessionPhoto } = sessionPhotoWithProject;
+    return sessionPhoto;
+  }
 
   async create(
     sessionId: string,
@@ -27,17 +67,15 @@ export class SessionPhotoService {
   }
 
   async findOne(id: string): Promise<GetSessionPhotoDto | null> {
-    return await this.prisma.sessionPhoto.findUnique({
-      where: {
-        id: id,
-      },
-    });
+    return this.authorize(id);
   }
 
   async update(
     id: string,
     updateSessionPhotoDto: UpdateSessionPhotoDto,
   ): Promise<void> {
+    this.authorize(id);
+
     await this.prisma.user.update({
       where: {
         id: id,
@@ -49,6 +87,8 @@ export class SessionPhotoService {
   }
 
   async remove(id: string): Promise<void> {
+    this.authorize(id);
+
     await this.prisma.sessionPhoto.delete({
       where: { id: id },
     });
